refactor: Simplify Card, CTA and extract component
