+++ nextjs/src/app/[locale]/theme/page.tsx
@@ -17,12 +17,15 @@ import {
 } from "./texts"
 import Container from "@/components/container"
 import CardSlider from "@/components/cards/card-slider"
+import CardSliderElement from "@/components/cards/card-slider-element"
 import SectionCardWide from "@/components/sections/section-card-wide"
 import SectionGroup from "@/components/sections/section-group"
 import LogoGroup from "@/components/logo-group"
 import ButtonGroup from "@/components/button-group"
 import CardIcon from "@/components/cards/card-icon"
 import CardIconGroup from "@/components/cards/card-icon-group"
+import CardColored from "@/components/cards/card-colored"
+
 export default function Theme() {
   const handleClick = () => {
     console.log("Button clicked!")
@@ -50,9 +53,20 @@ export default function Theme() {
         <CardSlider
           title={solutions.title}
           description={solutions.description}
-          cards={solutions.cards}
           ctas={solutions.ctas}
-        />
+        >
+          {solutions.cards.map((card, index) => {
+            return (
+              <CardSliderElement key={index}>
+                <CardColored
+                  color={card.color}
+                  title={card.title}
+                  description={card.description}
+                />
+              </CardSliderElement>
+            )
+          })}
+        </CardSlider>
       </Container>
 
       <Container background="stone" padding="both-padding">
