+++ nextjs/src/app/[locale]/theme/texts.ts
@@ -1,5 +1,5 @@
 import { type Card } from "@/components/cards/card"
-import { CTA } from "@/lib/actions"
+import { CTA } from "@/lib/cta"
 
 export const hero = {
   title: `## Potential. **Unlocked.**`,
@@ -65,7 +65,7 @@ export const cards = [
       "Lorem ipsum dolor sit amet, consec tetur adipiscing elit. Ut et massa mi. Aliquam in hendrerit urna. Pellentesque sit amet sapien fringilla, mattis ligula consectetur, ultrices mauris.",
     color: "sunglow",
   },
-] satisfies Card[]
+] as const
 
 const ctas = [
   {
