+++ nextjs/src/components/cards/card.ts
@@ -1,8 +1,5 @@
-import { colors } from "@/lib/colors"
-
 export interface Card {
   title: string
   description: string
-  color?: keyof typeof colors
   imageUrl?: string
 }
