+++ nextjs/src/components/sections/section-card-wide.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import CardImage from "../cards/card-image"
-import { CTA } from "@/lib/actions"
+import { CTA } from "@/lib/cta"
 import Link from "../link"
 import ButtonGroup from "../button-group"
 import clsx from "clsx"
